rpc_manager/requirements.txt
Django==2.1.2
grpcio==1.15.0
grpcio-tools==1.15.0
mysqlclient==1.3.12
protobuf==3.6.1
pytz==2018.5
six==1.11.0
uuid==1.30
